+++ smd/strings.py
@@ -1,4 +1,4 @@
-VERSION = "4.5.1"
+VERSION = "4.5.2"
 STEAM_WEB_API_KEY = "1DD0450A99F573693CD031EBB160907D"
 """NOTE: Not even mine lmao. Given away by oureveryday
 (source: https://github.com/SteamAutoCracks/Steam-auto-crack/issues/33)"""
@@ -7,5 +7,6 @@ REPO_NAME = "smd"
 # Update check source: https://github.com/Midrags/SMD_2/releases/
 GITHUB_UPDATE_USERNAME = "Midrags"
 REPO_UPDATE_NAME = "SMD_2"
+RELEASE_PAGE_URL = f"https://github.com/{GITHUB_UPDATE_USERNAME}/{REPO_UPDATE_NAME}/releases"
 WINDOWS_RELEASE_PREFIX = "0_windows_x86-64"
 LINUX_RELEASE_PREFIX = "1_linux_x86-64"